refactor: merge dtos into a file
